+++ packages/swirl-components/src/components/swirl-app-layout/swirl-app-layout.tsx
@@ -248,6 +248,10 @@ export class SwirlAppLayout {
       this.el.querySelector('[slot="app-bar-controls"]')
     );
 
+    const hasAppBarMobileMenuButton = Boolean(
+      this.el.querySelector('[slot="app-bar-mobile-menu-button"]')
+    );
+
     const className = classnames(
       "app-layout",
       `app-layout--mobile-view-${this.mobileView}`,
@@ -255,6 +259,7 @@ export class SwirlAppLayout {
       `app-layout--transitioning-to-${this.transitioningTo}`,
       `app-layout--transition-style-${this.transitionStyle}`,
       {
+        "app-layout--has-app-bar-mobile-menu-button": hasAppBarMobileMenuButton,
         "app-layout--has-app-bar-controls": hasAppBarControls,
         "app-layout--has-navigation": this.hasNavigation,
         "app-layout--has-sidebar": this.hasSidebar,