+++ packages/swirl-components/src/components/swirl-description-list-item/swirl-description-list-item.spec.tsx
@@ -12,10 +12,15 @@ describe("swirl-description-list-item", () => {
     expect(page.root).toEqualHtml(`
       <swirl-description-list-item role="listitem" term="Term">
         <mock:shadow-root>
-          <div class="description-list-item description-list-item--bordered description-list-item--orientation-horizontal" part="description-list-item" role="group">
-            <div class="description-list-item__term" part="description-list-item__term" role="term">Term</div>
-            <div class="description-list-item__description" role="definition">
-              <slot></slot>
+          <div class="description-list-item description-list-item--bordered" part="description-list-item" role="group">
+            <div class="description-list-item__text-container">
+              <div class="description-list-item__term" part="description-list-item__term" role="term">Term</div>
+              <div class="description-list-item__description" role="definition">
+                <slot></slot>
+              </div>
+            </div>
+            <div class="description-list-item__tools">
+              <slot name="tools"></slot>
             </div>
           </div>
         </mock:shadow-root>