+++ packages/swirl-components/templates.mjs
@@ -88,10 +88,10 @@ export class SwirlEmoji{{emojiName}} {
 
     return (
       <Fragment>
-        {this.size === 16 && <img alt={this.label} class={className} height="16" src="/emojis/{{emojiName}}16.png" width="16" />}
-        {this.size === 20 && <img alt={this.label} class={className} height="20" src="/emojis/{{emojiName}}20.png" width="20" />}
-        {this.size === 24 && <img alt={this.label} class={className} height="24" src="/emojis/{{emojiName}}24.png" width="24" />}
-        {this.size === 32 && <img alt={this.label} class={className} height="32" src="/emojis/{{emojiName}}32.png" width="32" />}
+        {this.size === 16 && <img alt={this.label} class={className} height="16" src="emojis/{{emojiName}}16.png" width="16" />}
+        {this.size === 20 && <img alt={this.label} class={className} height="20" src="emojis/{{emojiName}}20.png" width="20" />}
+        {this.size === 24 && <img alt={this.label} class={className} height="24" src="emojis/{{emojiName}}24.png" width="24" />}
+        {this.size === 32 && <img alt={this.label} class={className} height="32" src="emojis/{{emojiName}}32.png" width="32" />}
       </Fragment>
     );
   }