+++ packages/swirl-tokens/dart/lib/styles.light.dart
@@ -4,11 +4,7 @@
 //
 
 // Do not edit directly
-<<<<<<< HEAD
-// Generated on Tue, 29 Aug 2023 10:01:25 GMT
-=======
-// Generated on Mon, 09 Oct 2023 10:13:22 GMT
->>>>>>> main
+// Generated on Tue, 24 Oct 2023 12:24:55 GMT